+++ source/lib/windows/lib_init.cpp
@@ -14,16 +14,23 @@
 namespace ze_lib
 {
 #ifndef DYNAMIC_LOAD_LOADER
-    extern "C" BOOL APIENTRY DllMain(HINSTANCE hinstDLL, DWORD fdwReason, LPVOID lpvReserved) {
-        if (fdwReason == DLL_PROCESS_DETACH) {
-            delete context;
+    void context_dynamic_at_exit_destructor()
+    {
+        if (ze_lib::context) {
+            delete ze_lib::context;
+            ze_lib::context = nullptr;
+        }
+        if (loader::context) {
             delete loader::context;
-        } else if (fdwReason == DLL_PROCESS_ATTACH) {
+            loader::context = nullptr;
+        }
+    }
+    extern "C" BOOL APIENTRY DllMain(HINSTANCE hinstDLL, DWORD fdwReason, LPVOID lpvReserved) {
+        if (fdwReason == DLL_PROCESS_ATTACH) {
             context = new context_t;
             loader::context = new loader::context_t;
         }
         return TRUE;
-    }      
-
+    }
 #endif
 }